requirements.txt
GitPython>=2.1.11
